requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
